requirements.txt
locust==2.19.1
flask==3.0.0
prometheus_client
stomper
websockets
websocket_client
pycryptodome
google
protobuf
kafka-python
pymysql
